[PROT-460] Fixing CRM Task Issues

+++ crm/src/main/java/io/protone/crm/service/CrmOpportunityService.java
@@ -74,9 +74,13 @@ public class CrmOpportunityService {
 
     public void deleteOpportunityTask(String shortcut, Long taskId, String corNetwork) {
         CrmOpportunity crmOpportunity = opportunityRepository.findOneByShortNameAndNetwork_Shortcut(shortcut, corNetwork);
-        crmOpportunity.getTasks().removeIf(crmTask -> crmTask.getId() == taskId);
-        opportunityRepository.save(crmOpportunity);
-        crmTaskService.deleteByIdAndNetwork_Shortcut(taskId, corNetwork);
+        CrmTask crmTask = crmTaskService.findOneByIdAndNetwork_Shortcut(taskId, corNetwork);
+        if (crmTask != null) {
+            crmOpportunity.removeTasks(crmTask);
+            crmTaskService.saveOrUpdateTaskAssociatiedWithOpportunity(crmOpportunity, crmTask);
+            opportunityRepository.saveAndFlush(crmOpportunity);
+            crmTaskService.deleteByIdAndNetwork_Shortcut(taskId, corNetwork);
+        }
     }
